requirements.txt
ml-dtypes>=0.4.0,<0.5.0
protobuf>=3.20.3,<6.0.0
tensorflow-intel==2.18.0
tensorflow==2.18.0
keras==3.7.0
numpy==2.0.2
pillow==11.0.0
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.36
tensorboard==2.18.0
requests==2.32.3